Moves DataTransformation components into a new subfolder, breaks a couple new components out, adds the final data transformation (adjust geo) to the new custom hook based flow
