+++ ui/client/datasets/DataTransformation/ClipMap.js
