+++ ui/client/datasets/DataTransformation/useElwoodData.js
@@ -1,0 +1,104 @@
+import { useState, useEffect } from 'react';
+import axios from 'axios';
+
+// Hook that handles all the data fetching from Elwood
+const useElwoodData = ({
+  datasetId,
+  annotations,
+  onSuccess,
+  generateArgs,
+  jobString,
+  cleanupRef,
+}) => {
+  const [data, setData] = useState(null);
+  const [dataLoading, setDataLoading] = useState(false);
+  const [dataError, setDataError] = useState(false);
+
+  useEffect(() => {
+    const startElwoodJob = async ({ requestArgs }) => {
+      console.log('these are the requestArgs when starting ', jobString, requestArgs)
+      const jobQueueResp = await axios.post(
+        `/api/dojo/job/${datasetId}/${jobString}`, requestArgs
+      );
+
+      if (jobQueueResp.status === 200) {
+        return jobQueueResp.data?.id;
+      }
+    };
+
+    const runElwoodJob = ({ requestArgs, onFailure }) => {
+      console.log('runElwoodJob has been called at the top level');
+      console.log('this is requestArgs in ', jobString, requestArgs)
+      let count = 0;
+
+      const repeatFetch = (jobId) => {
+        setTimeout(() => {
+          axios.post(`/api/dojo/job/fetch/${jobId}`).then((response) => {
+            if (response.status === 200) {
+              // keep track of how long it takes (for dev purposes)
+              count += 1;
+              console.log(`${count}: response from job string: ${jobString}:`, response);
+              if (response.data) {
+                console.log(`success! it took ${count * 500}ms`, response.data);
+                onSuccess(response.data, setData, setDataError, setDataLoading);
+                return;
+              }
+              if (cleanupRef.current) {
+                // if no data and the component is mounted, try the fetch again
+                repeatFetch(jobId);
+              }
+            }
+          }).catch(() => {
+            // we get a 404 immediately - there is some sort of bug, this accounts for that
+            if (count < 2 && cleanupRef.current) {
+              // and the component is mounted
+              console.log('repeating job call');
+              startElwoodJob({ requestArgs }).then((resp) => {
+                repeatFetch(resp);
+              });
+            } else {
+              onFailure();
+            }
+          });
+        }, 500);
+      };
+
+      if (datasetId) {
+        startElwoodJob({ requestArgs }).then((jobId) => repeatFetch(jobId));
+      }
+    };
+
+    if (!data && !dataError && !dataLoading) {
+      if (Object.keys(annotations?.annotations).length) {
+        // don't do any of the below until annotations has been populated
+        setDataLoading(true);
+        const args = generateArgs(annotations);
+        const onFailure = () => {
+          setDataError(true);
+          setDataLoading(false);
+        };
+        if (Object.keys(args).length) {
+          runElwoodJob({
+            requestArgs: args,
+            onFailure,
+          });
+        }
+      }
+    }
+    // The linter is complaining about cleanupRef missing here because it doesn't understand
+    // that this is a reference to a ref, which both won't and shouldn't trigger a rerender
+  }, [
+    datasetId,
+    annotations,
+    jobString,
+    data,
+    dataError,
+    dataLoading,
+    generateArgs,
+    onSuccess,
+  ]);
+
+  return { data, error: dataError };
+};
+
+export default useElwoodData;
